+++ src/components/body/slideshow/index.js
@@ -1,3 +1,4 @@
+import Img from '../../img-with-placeholder'
 import Multimedia from '../multimedia'
 import NextArrowSvg from '../../../assets/body/slideshow/next-arrow.svg'
 import PreArrowSvg from '../../../assets/body/slideshow/pre-arrow.svg'
@@ -5,7 +6,8 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import get from 'lodash/get'
 import map from 'lodash/map'
-import mq from '../../../utils/media-query'
+import memoize from 'memoize-one'
+import mq from '@twreporter/core/lib/utils/media-query'
 import styled from 'styled-components'
 
 const _ = {
@@ -518,22 +520,6 @@ export default class Slideshow extends PureComponent {
 
   constructor(props) {
     super(props)
-    const images = _.get(props, 'data.content', [])
-    this.total = images.length
-
-    // For slicing images array easily later,
-    // add last `slidesOffset` elements into top of images array.
-    // add first `slidesOffset` elements into bottom of images array.
-    // EX:
-    // slidesOffset: 2
-    // input images: [ a, b, c, d ]
-    // output images: [c, d, a, b, c, d, a, b]
-    this.images = [].concat(
-      images.slice(-slidesOffset),
-      images,
-      images.slice(defaultCurIndex, slidesOffset)
-    )
-
     this.defaultTranslateXUnit = -slidesOffset
 
     this.state = {
@@ -549,6 +535,7 @@ export default class Slideshow extends PureComponent {
   }
 
   _slideToPrev() {
+    const total = _.get(this.props, 'data.content.length', 0)
     this.setState(
       {
         isSliding: true,
@@ -558,7 +545,7 @@ export default class Slideshow extends PureComponent {
         let curSlideIndex = this.state.curSlideIndex - 1
 
         if (curSlideIndex < defaultCurIndex) {
-          curSlideIndex = this.total + curSlideIndex
+          curSlideIndex = total + curSlideIndex
         }
 
         setTimeout(() => {
@@ -573,6 +560,7 @@ export default class Slideshow extends PureComponent {
   }
 
   _slideToNext() {
+    const total = _.get(this.props, 'data.content.length', 0)
     this.setState(
       {
         isSliding: true,
@@ -581,8 +569,8 @@ export default class Slideshow extends PureComponent {
       () => {
         let curSlideIndex = this.state.curSlideIndex + 1
 
-        if (curSlideIndex >= this.total) {
-          curSlideIndex = curSlideIndex % this.total
+        if (curSlideIndex >= total) {
+          curSlideIndex = curSlideIndex % total
         }
 
         setTimeout(() => {
@@ -596,32 +584,51 @@ export default class Slideshow extends PureComponent {
     )
   }
 
+  buildImagesForSlicing = memoize(images => {
+    // For slicing images array easily later,
+    // add last `slidesOffset` elements into top of images array.
+    // add first `slidesOffset` elements into bottom of images array.
+    // EX:
+    // slidesOffset: 2
+    // input images: [ a, b, c, d ]
+    // output images: [c, d, a, b, c, d, a, b]
+    return [].concat(
+      images.slice(-slidesOffset),
+      images,
+      images.slice(defaultCurIndex, slidesOffset)
+    )
+  })
+
   render() {
     const { curSlideIndex, isSliding, translateXUint } = this.state
+    const images = _.get(this.props, 'data.content', [])
+    const total = images.length
 
-    const slides = this.images.slice(
+    const imagesForSlicing = this.buildImagesForSlicing(images)
+
+    const slides = imagesForSlicing.slice(
       curSlideIndex,
       curSlideIndex + slidesOffset * 2 + 1
     )
 
-    const slidesJSX = _.map(slides, (slide, index) => {
+    const slidesJSX = _.map(slides, (slide = {}, index) => {
+      const objectFit =
+        _.get(slide, 'mobile.width', 0) > _.get(slide, 'mobile.height', 0)
+          ? 'cover'
+          : 'contain'
       return (
         <SlideFlexItem key={`slide_${slide.id}`}>
-          {/* TODO use ImgWrapper */}
-          <img
-            src={slide.tablet.url}
-            width="100%"
-            height="100%"
-            style={{
-              objectFit:
-                slide.tablet.width > slide.tablet.height ? 'cover' : 'contain',
-            }}
+          <Img
+            imageSet={[slide.mobile, slide.tablet, slide.desktop]}
+            defaultImage={slide.mobile}
+            objectFit={objectFit}
+            sizes="(max-width: 800px) 800px, (max-width: 1200px) 1200px, 2000px"
           />
         </SlideFlexItem>
       )
     })
 
-    const desc = _.get(slides, [curSlideIndex, 'description'])
+    const desc = _.get(images, [curSlideIndex, 'description'])
 
     return (
       <SlideshowFlexBox>
@@ -646,7 +653,7 @@ export default class Slideshow extends PureComponent {
         </PrevNextSection>
         <ImageNumberCircle>
           <ImageNumber>{curSlideIndex + 1}</ImageNumber>
-          <ImageTotal>{this.total}</ImageTotal>
+          <ImageTotal>{total}</ImageTotal>
         </ImageNumberCircle>
         {desc ? <Desc>{desc}</Desc> : <EmptyDesc />}
       </SlideshowFlexBox>