+++ CHANGELOG.md
@@ -10,7 +10,26 @@
 
 - Provide `primary`, `secondary` and `base` colors for default theme
 - Support `extend`, `normal`, `fullscreen` and `small` leading block with default theme
+
+#### Photo Theme(article:v2:photo) Support
+
+- Provide `primary`, `secondary` and `base` colors for photo theme
+- Support `extend`, `normal`, `fullscreen` and `small` leading block with photo theme
+
+#### UIManager
+
 - Introduce `UIManager(src/managers/ui-manager.js)` to handle theme and layout
+
+#### Bug fix
+
+- Fix body:slideshow bug: description not rendered if current slide index is more than 6
+- Fix body:centered-quote bug: `by` -> `quoteBy`
+
+#### Miscellaneous
+
+- URL origin replacement: storage.googleapis.com -> www.twreporter.org
+- body:slideshow updated: img tag -> img-with-placeholder
+- src/utils/media-query -> @twreporter/core/lib/utils/media-query
 
 ### 1.0.5